+++ src/LiquidPixel.java
@@ -22,7 +22,7 @@ public abstract class LiquidPixel extends Pixel {
 	 * @param freezePoint The freezing point of the pixel
 	 * @param vapPoint    The vaporization point of the pixel
 	 */
-	LiquidPixel(double density, double temp, double freezePoint, double vapPoint, int r, int g, int b, int a) {
+	LiquidPixel(double density, double temp, double freezePoint, double vapPoint, int r, int g, int b, double a) {
 		super(PixelType.LIQUID, density, temp, r, g, b, a);
 		this.freezePoint = freezePoint;
 		this.vapPoint = vapPoint;
